Move .all() definition and tests into separate directories
