+++ methods/all.js
